+++ src/main/java/saboroso/saborosoburguer/repositories/AddOnRepository.java
@@ -2,7 +2,7 @@ package saboroso.saborosoburguer.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import saboroso.saborosoburguer.entities.menuItems.AddOn;
+import saboroso.saborosoburguer.entities.menuItems.accompaniment.AddOn;
 
 import java.util.List;
