+++ src/main/java/saboroso/saborosoburguer/repositories/AddOnSaleRepository.java
@@ -1,7 +1,7 @@
 package saboroso.saborosoburguer.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import saboroso.saborosoburguer.entities.soldItems.AddOnSale;
+import saboroso.saborosoburguer.entities.soldItems.accompaniment.AddOnSale;
 
 public interface AddOnSaleRepository extends JpaRepository<AddOnSale, Long> {
 }
