+++ src/main/java/saboroso/saborosoburguer/repositories/ComboRepository.java
@@ -2,7 +2,7 @@ package saboroso.saborosoburguer.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import saboroso.saborosoburguer.entities.menuItems.Combo;
+import saboroso.saborosoburguer.entities.menuItems.accompaniment.Combo;
 
 import java.util.List;
