+++ src/main/java/saboroso/saborosoburguer/repositories/ComboSaleRepository.java
@@ -1,7 +1,7 @@
 package saboroso.saborosoburguer.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import saboroso.saborosoburguer.entities.soldItems.ComboSale;
+import saboroso.saborosoburguer.entities.soldItems.accompaniment.ComboSale;
 
 public interface ComboSaleRepository extends JpaRepository <ComboSale, Long> {
 }
