+++ src/main/java/saboroso/saborosoburguer/services/ComboService.java
@@ -3,7 +3,7 @@ package saboroso.saborosoburguer.services;
 import org.springframework.stereotype.Service;
 import saboroso.saborosoburguer.DTO.combo.ComboDTO;
 import saboroso.saborosoburguer.DTO.combo.ComboMapper;
-import saboroso.saborosoburguer.entities.menuItems.Combo;
+import saboroso.saborosoburguer.entities.menuItems.accompaniment.Combo;
 import saboroso.saborosoburguer.models.CRUDResponseMessage;
 import saboroso.saborosoburguer.repositories.ComboRepository;
